Move redux into components
