+++ src/components/app/app.reducers.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import auth from './auth'
+import auth from '../login/login.reducers'
 import { routerReducer } from 'react-router-redux';
 
 export default combineReducers({ auth, router: routerReducer });
