+++ src/components/login/login.types.js
@@ -1,0 +1,1 @@
+export const LOGIN_USER = "LOGIN_USER";
