--- src/redux/actions/actionTypes.js
@@ -1,1 +1,0 @@
-export const LOGIN_USER = "LOGIN_USER";
